+++ src/ng-redux.module.ts
@@ -1,15 +1,15 @@
-import { NgModule } from '@angular/core';
+import { NgModule, NgZone } from '@angular/core';
 import { NgRedux } from './components/ng-redux';
 import { DevToolsExtension } from './components/dev-tools';
 
-export function _ngReduxFactory() {
-  return new NgRedux();
+export function _ngReduxFactory(ngZone: NgZone) {
+  return new NgRedux(ngZone);
 }
 
 @NgModule({
   providers: [
     DevToolsExtension,
-    { provide: NgRedux, useFactory: _ngReduxFactory }
+    { provide: NgRedux, useFactory: _ngReduxFactory, deps: [ NgZone ] }
   ]
 })
 export class NgReduxModule { };